+++ c/qrcodegen-test.c
@@ -114,13 +114,13 @@ static uint8_t *addEccAndInterleaveReference(const uint8_t *data, int version, e
 	int shortBlockLen = rawCodewords / numBlocks;
 	
 	// Split data into blocks and append ECC to each block
-	uint8_t **blocks = malloc(numBlocks * sizeof(uint8_t*));
-	uint8_t *generator = malloc(blockEccLen * sizeof(uint8_t));
+	uint8_t **blocks = malloc((size_t)numBlocks * sizeof(uint8_t*));
+	uint8_t *generator = malloc((size_t)blockEccLen * sizeof(uint8_t));
 	reedSolomonComputeDivisor(blockEccLen, generator);
 	for (int i = 0, k = 0; i < numBlocks; i++) {
-		uint8_t *block = malloc((shortBlockLen + 1) * sizeof(uint8_t));
+		uint8_t *block = malloc((size_t)(shortBlockLen + 1) * sizeof(uint8_t));
 		int datLen = shortBlockLen - blockEccLen + (i < numShortBlocks ? 0 : 1);
-		memcpy(block, &data[k], datLen * sizeof(uint8_t));
+		memcpy(block, &data[k], (size_t)datLen * sizeof(uint8_t));
 		reedSolomonComputeRemainder(&data[k], datLen, generator, blockEccLen, &block[shortBlockLen + 1 - blockEccLen]);
 		k += datLen;
 		blocks[i] = block;
@@ -128,7 +128,7 @@ static uint8_t *addEccAndInterleaveReference(const uint8_t *data, int version, e
 	free(generator);
 	
 	// Interleave (not concatenate) the bytes from every block into a single sequence
-	uint8_t *result = malloc(rawCodewords * sizeof(uint8_t));
+	uint8_t *result = malloc((size_t)rawCodewords * sizeof(uint8_t));
 	for (int i = 0, k = 0; i < shortBlockLen + 1; i++) {
 		for (int j = 0; j < numBlocks; j++) {
 			// Skip the padding byte in short blocks
@@ -149,18 +149,18 @@ static void testAddEccAndInterleave(void) {
 	for (int version = 1; version <= 40; version++) {
 		for (int ecl = 0; ecl < 4; ecl++) {
 			int dataLen = getNumDataCodewords(version, (enum qrcodegen_Ecc)ecl);
-			uint8_t *pureData = malloc(dataLen * sizeof(uint8_t));
+			uint8_t *pureData = malloc((size_t)dataLen * sizeof(uint8_t));
 			for (int i = 0; i < dataLen; i++)
-				pureData[i] = rand() % 256;
+				pureData[i] = (uint8_t)(rand() % 256);
 			uint8_t *expectOutput = addEccAndInterleaveReference(pureData, version, (enum qrcodegen_Ecc)ecl);
 			
 			int dataAndEccLen = getNumRawDataModules(version) / 8;
-			uint8_t *paddedData = malloc(dataAndEccLen * sizeof(uint8_t));
-			memcpy(paddedData, pureData, dataLen * sizeof(uint8_t));
-			uint8_t *actualOutput = malloc(dataAndEccLen * sizeof(uint8_t));
+			uint8_t *paddedData = malloc((size_t)dataAndEccLen * sizeof(uint8_t));
+			memcpy(paddedData, pureData, (size_t)dataLen * sizeof(uint8_t));
+			uint8_t *actualOutput = malloc((size_t)dataAndEccLen * sizeof(uint8_t));
 			addEccAndInterleave(paddedData, version, (enum qrcodegen_Ecc)ecl, actualOutput);
 			
-			assert(memcmp(actualOutput, expectOutput, dataAndEccLen * sizeof(uint8_t)) == 0);
+			assert(memcmp(actualOutput, expectOutput, (size_t)dataAndEccLen * sizeof(uint8_t)) == 0);
 			free(pureData);
 			free(expectOutput);
 			free(paddedData);
@@ -362,7 +362,7 @@ static void testReedSolomonMultiply(void) {
 
 static void testInitializeFunctionModulesEtc(void) {
 	for (int ver = 1; ver <= 40; ver++) {
-		uint8_t *qrcode = malloc(qrcodegen_BUFFER_LEN_FOR_VERSION(ver) * sizeof(uint8_t));
+		uint8_t *qrcode = malloc((size_t)qrcodegen_BUFFER_LEN_FOR_VERSION(ver) * sizeof(uint8_t));
 		assert(qrcode != NULL);
 		initializeFunctionModules(ver, qrcode);
 		
@@ -718,7 +718,7 @@ static void testCalcSegmentBitLength(void) {
 			{INT_MAX / 1, -1},
 		};
 		for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
-			assert(calcSegmentBitLength(qrcodegen_Mode_NUMERIC, cases[i][0]) == cases[i][1]);
+			assert(calcSegmentBitLength(qrcodegen_Mode_NUMERIC, (size_t)cases[i][0]) == cases[i][1]);
 			numTestCases++;
 		}
 	}
@@ -748,7 +748,7 @@ static void testCalcSegmentBitLength(void) {
 			{INT_MAX / 1, -1},
 		};
 		for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
-			assert(calcSegmentBitLength(qrcodegen_Mode_ALPHANUMERIC, cases[i][0]) == cases[i][1]);
+			assert(calcSegmentBitLength(qrcodegen_Mode_ALPHANUMERIC, (size_t)cases[i][0]) == cases[i][1]);
 			numTestCases++;
 		}
 	}
@@ -777,7 +777,7 @@ static void testCalcSegmentBitLength(void) {
 			{INT_MAX / 1, -1},
 		};
 		for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
-			assert(calcSegmentBitLength(qrcodegen_Mode_BYTE, cases[i][0]) == cases[i][1]);
+			assert(calcSegmentBitLength(qrcodegen_Mode_BYTE, (size_t)cases[i][0]) == cases[i][1]);
 			numTestCases++;
 		}
 	}
@@ -805,7 +805,7 @@ static void testCalcSegmentBitLength(void) {
 			{INT_MAX / 1, -1},
 		};
 		for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
-			assert(calcSegmentBitLength(qrcodegen_Mode_KANJI, cases[i][0]) == cases[i][1]);
+			assert(calcSegmentBitLength(qrcodegen_Mode_KANJI, (size_t)cases[i][0]) == cases[i][1]);
 			numTestCases++;
 		}
 	}
@@ -1049,7 +1049,7 @@ static void testGetTotalBits(void) {
 /*---- Main runner ----*/
 
 int main(void) {
-	srand(time(NULL));
+	srand((unsigned int)time(NULL));
 	testAppendBitsToBuffer();
 	testAddEccAndInterleave();
 	testGetNumDataCodewords();
